[AV-1691] Refactor sync blocks (#786)
* Refactor sync blocks

+++ consensus/dummy/consensus.go
@@ -199,7 +199,7 @@ func (self *DummyEngine) verifyHeader(chain consensus.ChainHeaderReader, header 
 			return fmt.Errorf("extra-data too long: %d > %d", len(header.Extra), params.MaximumExtraDataSize)
 		}
 	} else {
-		if len(header.Extra) != params.ApricotPhase3ExtraDataSize {
+		if uint64(len(header.Extra)) != params.ApricotPhase3ExtraDataSize {
 			return fmt.Errorf("expected extra-data field to be: %d, but found %d", params.ApricotPhase3ExtraDataSize, len(header.Extra))
 		}
 	}
